requirements.txt
# VMware VM Monitor Dependencies
# This project primarily uses Python standard library and system tools

# No external dependencies required for basic functionality
# The project uses:
# - subprocess (built-in) for vmrun commands
# - logging (built-in) for logging
# - datetime (built-in) for time operations
# - os (built-in) for file operations
# - tempfile (built-in) for temporary files

# Optional dependencies for enhanced functionality:

# For advanced Windows event log parsing (optional)
# pywin32>=227

# For configuration file support (optional)
pyyaml>=6.0

# For enhanced logging (optional)
# colorlog>=6.7.0

# For process monitoring (optional)
# psutil>=5.9.0

# For email notifications (optional)
# smtplib is built-in, but for advanced email features:
# email-validator>=1.3.0

# For web dashboard (optional)
# flask>=2.3.0
# flask-socketio>=5.3.0

# Development dependencies (uncomment if needed)
# pytest>=7.4.0
# pytest-cov>=4.1.0
# black>=23.7.0
# flake8>=6.0.0